mzcompose: Allow complete control over docker-compose env vars
For some upcoming AWS testing it is necessary to pass AWS env vars to mzcompose
and testdrive but not materialized. Here I expose subprocess's ability to block
variables that mzcompose has access to to composition files.

There was only one use of the existing `env` option, so I ported it over to use
the new structure.