coord: do not produce timestamps in advance of the clock
Previously if `INSERT` and `SELECT` statements on tables were interleaved
and executed as fast as possible, our timestamp code code produce
timestamps increasingly in advance of `now()`. This happened because
if the clock ever returned a timestamp that wasn't greater than the
previous value, it would add 1. Take some case where a system clock
retreated many seconds. New table data continue to be inserted in the
future until (which might never happen) the system clock catches up with
the timestamps that we have been forcing forward.

Instead, allow the function that produces write timestamps for tables
to fail. In this case, callers of that function can choose what to
do. Internal metrics drop their updates. User inserts block, queing
themselves for retry once there is a new timestamp produced by the
system clock in advance of anything it has done before. Reads at this
time still function (but don't observe the queued inserts). Achieve this
by creating an enum to describe this situation.

Fixes #8424