Use the default toolchain.
This causes developers to use the default toolchain (normally `stable`),
rather than pinning to a specific named Rust version. We add a minimum
rust version specifier to each crate, to enforce that developers
are not building with older versions, and add a linter to enforce
that these are kept in sync.

This has no effect on the toolchain used to build in CI.