dataflow: use ProbeHandles to communicate source uppers
Previously, we only communicated source uppers from the dataflow crate to the
coordinator for a small subset of sources which had real time timestamp bindings.
The uppers which we did communicate in that case corresponded to the upper frontier
of known timestamp bindings ie the upper frontier for timestamps which we currently
knew how to assign data for given a source like Kafka topic.

This doesn't correspond to the concept of "what times are ready to be queried now",
as the timestamps had not even been made definite yet, and it was unclear how to
extend this approach to CDCv2 / byo sources that did not have timestamp bindings.

This commit changes the behavior to instead:

- define a source's upper as "the maximum time in any source instance that has
  fully passed through the 'source pipeline'", where the source pipeline will
  be responsible for making the source definite.
- Detect that frontier using ProbeHandles on each source instance.