fixup! ignore test failure on csv sources + move probe to storage
This commit does two things + a small refactor:

1. It moves the ProbeHandle to the earliest point at which there is a differential
collection that can reasonably be considered definite.

2. It ignores a test failure that happens for csv sources.

Basically, previously, even when a csv source was closed, we would still generate
new RT bindings for it, and continue to compact those bindings. In other words,
even though the source was closed we would still present a non-empty upper for
the source, and use that non-empty upper to drive source compaction. Now, we correctly
present an empty upper when the source is closed. However, coord declines to compact
up to the empty upper, as doing so makes the data unusable. The old real-time timestamping
code continues to send new timestamp bindings, but those bindings are now never compacted,
because the coordinator never advances the since frontier because upper is empty.

Ideally, the real time timestamping code would just stop creating new timestamp
bindings when the source is permanently closed, but that abstraction doesn’t
exist yet.

Finally, for the small refactor - this commit changes some of the variable
names in import_source to make it more clear that these variables are Collections
and not Streams.