test: fix test for cdcv2
Previously, this test tried to emulate a produce belatedly sending a stream of
data that CDCv2 should ignore, but it did not send exact copies of ((k, v) t, d)
as expected and instead forwarded the times to a later time. The test
also had two instances of the same ((k, v), t) in a batch which is not valid
CDCv2 input.

This commit fixes the test to emit actually previously emitted data that obeys
the requirement that CDCv2 batches are consolidated.