mzbuild: perform inline graph traversal when acquiring deps
Per @petrosagg's code review on #10267, mzbuild gets much simpler if we
memoize dependency acquisition and perform a much simpler graph
traversal inline in `ResolvedImage.acquire`.

The one casualty here is the `force_build` parameter to
DependencySet.acquire. That existed only to support `mzimage build`,
which was not particularly important. We could add this back with not
too much effort, but didn't seem worth it at the moment.