mzcompose: Support specifying volumes in py files
Currently if you want to specify volumes you are required to add an
mzcompose.yml file next to mzcompose.py.

With this, VOLUMES in mzcompose.py acts analogously to SERVICES, and
VOLUMES_EXTRA adds volumes in addition to the default volumes, but does not
work along with any other explicitly-set volumes.