sql: support queries in TAIL
Implement by creating a temporary view then TAILing that. These views
are normal temp views and so are cleaned up when the session ends,
not when the transaction ends. This will slowly pollute the catalog
if connections are pooled and sessions are long lived. However this is
already what some users are doing, so it's fine for now.

Unexplored yet if it's possible to drop the views at txn end due to how
catalog_transact works (and thus could error).

Fixes #5484