doc/platform: flesh out user experience doc
Add quite a bit detail to the user experience document, so we can start
having focused design discussions.

The user experience document maps the available architectural components
to user-facing concepts. In particular, it proposes names for these
components and describes how end users will interact with them.

Sometimes the mapping is a direct consequence of technical restrictions,
and sometimes the mapping is flexible and the decision rests entirely on
the desired user experience. The "Discussion" section at the end of the
doc describes the thornier moments I encountered while developing this
mapping.