mzbuild,xcompile: Compile the coverage build using the nightly compiler
Coverage builds require the -Zinstrument-coverage rustc option, which
is only available in Nightly. So pass the desire to build a coverage
build all the way to where the bin/ci-builder command line is composed,
so that the nightly compiler is chosen.

Add the state of the coverage option to the hash that is used to decide
if an image needs to be rebuilt. Otherwise, a non-coverage image will
be used if already present in Docker.