scratch: add login subcommand
Add a `login` subcommand to `bin/scratch` that delegates to `aws sso
login` with the correct profile. This is a good bit easier to type than
`aws sso login --profile mz-scratch-admin`.

A minor refactoring was necessary to prevent argument parsing for the
`mine` subcommand from attempting to make a API call to AWS to figure
out the invoking user's email address.