sql: refactor list/array subscripting for PG compat
Subscript operations used to be left associative binary operations,
e.g.

LIST[[1,2],[3,4],[5,6]][1][2]

was treated as

(
  (
    LIST[[1,2],[3,4],[5,6]]
    [1]
  )
  [2]
)

However, PostgreSQL doesn't work like this, which is more obvious
in the case of slciing; to get PG's slicing semantics, we had
previously used a different, comma-separate syntax. However, all ww
actually needed to do was the refactor in this PR: collect all
subscript operations into a vec and pass them along to the subscript
func impl.

Though this PR doesn't include it, this work is necessary to support
subscripts on int2vector, for which multi-dimensional subscripts
should return NULL. In the prior implementation the second subscript
would successfully descend into the inner int2vector.