test: emit correct CDCv2 data in consolidation test
Using BYO sources we were able to emit multiple updates for a given
(data, time) pair and therefore test consolidation of records on the
same timestamp.

The CDCv2 protocol requires that a statement for a given (data, time)
pair is made only once and it's final, otherwise the wrong stream is
reproduced. Unfortunately this means that we can't test the case where
records consolidate within a single timestamp right after ingestion.

This isn't as bad though because the rest of the test enables compaction
which eventually does consolidate records and so we can test it through
there.

In some glorious feature we might end up with a special testdrive source
type that we simply type in the precise (Row, Timestamp, Diff) updates
verbatim instead of jumping through hoops with avro encoded CDCv2
sources.

Fixes #10397

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>