persist: split unsealed_drain into req/res components
This sets us up to run unsealed_drain in background maintenance, but the
immediate motivation is to unblock a PR that changes when listeners
fire.

To enable sending the relevant unsealed snapshot as part of a request,
an UnsealedSnapshotMeta is extracted. This has the same info as a
UnsealedSnapshot but with blob keys instead of get futures. A `fetch`
method on this converts it into the same UnsealedSnapshot we had before
by kicking off the fetches.