coord: hoist persist runtime out of catalog
Hoist the persist runtime out of the coordinator's catalog. This is a
bit delicate, because the cluster ID stored in the catalog is needed to
start the persist runtime. The new startup sequence looks like this:

  * Open the coordinator catalog.
  * Extract the catalog's cluster ID and use it to start the persist
    runtime.
  * Start the dataflow server, passing in the persist runtime.
  * Start the coordinator, passing in the persist runtime.

This tees us up to use separate persist runtimes in STORAGE and COMPUTE.
It also removes the irritating `StorageCommand::EnablePersistence`
variant, which was not serializable.

As part of this refactor, I moved table persistence handling to the
PersisterWithConfig struct, rather than making them part of the catalog.
This is part of our general effort to separate hard start from soft
state. It also brings tables in line with sources, which already store
only their hard state in the catalog.