[kafka exactly once sink] Reenable timestamp compaction verification (#10417)
Disabled in #10413 because it was flaky with a large number of workers.

Just increase the timeout and re-enable.  Flaking (as opposed to perma-failing) isn't really indicative of an issue since we just want the compaction to _eventually_ happen.  We're not trying to put strong bounds on how quickly timestamp compaction must occur after it's theoretically possible -- we just want to make sure it _does_ happen so the timestamp bindings database doesn't explode in size.  Note that we do allow _progress_ in compaction (as opposed to just looking for an absolute number) to satisfy verification.

The underlying issue was that the `permit-progress=true` flag wasn't on the failing verification because it was commented out when that flag was added

Fixes #10398