persist: add a limit to the blobcache size
Previously, the blob cache used by persist was unbounded which lead to unbounded
memory usage during ingestion. This commit adds a 2 GB limit to the cache, split
evenly between unsealed and trace. Objects are added from the cache iff adding
the object to the cache would not exceed the cache's size limit. Objects are
never evicted from the cache unless they are deleted by the user.

This design is counterintuitive, as LRU eviction is a commonly used cache management
policy. LRU works well if reads objects are expected to be inserted, read several times,
and then deleted and furthermore, reads have good temporal locality (i.e. an object
that was inserted recently / read recently has a good likelihood of being read
again soon).

Unfortunately, that is not the case for persist's current usage of blob. Breaking
things down a little bit more:

- Unsealed batches operate like a queue. Batches are only ever read from unsealed
  by the arrangement to remove them and translate them into a trace batch, in
  the order that they came in (assuming data arrives in roughly timestamp order).
  Therefore, evicting an old cached unsealed batch in favor of a newly inserted
  unsealed batch would be entirely useless. Similarly, caching an unsealed batch
  after the read to move data to trace is not helpful because that batch will be
  deleted soon.
- Trace batches operate not exactly like a queue but have some queue like
  properties. Traces compact batches going from oldest to most recent looking for
  adjacent pairs at the same compaction level. This means that in some situations
  (e.g. when theres a long line of trace batches at the same compaction level)
  removing the oldest cached batches is unproductive. Other times (e.g when theres
  a long lived trace batch at level say 20, but all of the other batches are at
  levels < 5, it makes sense to evict that level 20 batch because it will be a
  very long time before we can generate another level 20 batch. But without more
  fine grained input from the user, its not really clear what heuristic to use
  to detect this. The level 20 batch may not always be the oldest and it might
  not neccessarily be the largest either. Finally, arrangements read from the
  trace batches to drive compaction, but those reads have no temporal locality
  because after compacting the batches that were read are deleted.
- Snapshots currently happen very rarely (only on startup) and when they do,
  they read all batches so snapshots don't help us in implementing an LRU
  style heuristic (because after a snapshot every batch has the exact same
  last used time).

For all of these reasons, this commit does not teach the blob cache to do cache
evictions. This policy is likely also "good enough" because every batch in the
system gets deleted eventually, and no batch or set of batches will block others
from being cached forever.

Finally, this commit splits up separate quotas of memory usage for the trace and
unsealed batches because we hypothesize that:

1. Caching is likely more beneficial for unsealed because it is a more predictable
   workload and unsealed batches are likely smaller on average than their trace
   counterparts which get compacted together.
2. Trace batching is more likely to mess up and leave the cache in a less useful state
   where a few long lived trace batches prevent everything else from being cached
   for a long time.
3. In conclusion, keeping trace and unsealed caches physically separate seemed like
   a decent strategy to ensure we had a minimal amount of space reserved specifically
   for the unsealed use case. This obviously has the downside that perhaps some space
   sits empty when it could be utilized by trace batches.