testdrive: remove `set-execution-count` action
This action is unused, so remove it. It required a lot of complexity to
support, since it broke the testdrive action paradigm. The mzcompose
infrastructure on top of testdrive can be used to invoke testdrive
multiple times if required, or the same file can simply be provided on
the command line multiple times in a row.