dataflow, persist: don't use shared TimestampHistories to pass around restored offsets
Before, we were using the `TimsetampHistoriesBox`, which is shared
between all instances of a given source to pass around restored starting
offsets. We enfore that there can only be one rendered instance of a
source, but if we would render, then drop, and then render a sink again,
the previously set restored offsets would persist (ha!), meaning we
would try and restore the newly rendered source at the now outdated
starting offsets.

Now, we remove that functionality from `TimestampHistoriesBox/Rc` and
instead pass the restored offsets around in `create_source`.