persist: in persist() operator, prevent frontier from advancing once we error
Before, we would not hold on to a capability for data that we failed to
persist, meaning the frontier could advance in the case of write errors.
This meant that a downstream seal() operator could potentially advance
the seal, even though data that should have been written for a given
time was not written. This manifested in panics from the consistency
checks of, for example, the upsert() operator, which checks that its
restored updates sum up to a diff of 1.