persist: give "running" token to seal(), stop sealing when it's dropped
Before, spurious frontier changes could happen while a pipeline/dataflow
is being shut down. Now, we give a token to seal that allows it to
notice when the dataflow in which it is running is being shut down. When
the token is dropped, we don't react to changes in the input frontier by
advancing the seal frontier.

This would manifest in tests when errors happen in `persist()` (which
hold back the downstream frontier). In that case, the frontier would be
held, but on shutdown the `persist()` operator would drop all
capabilities, meaning the downstream frontier (at the `seal()` operator)
could advance. But reacting to that advancement would be wrong because
it would seal while there where write errors, meaning data would be
missing.