coord/persist: lookup tables with the correct table id, not the primary index id
Fixes #10533

We have to do a complicated dance at the moment to propagate since updates to persisted
collections for tables. We first grab the since updates for a given index, then check
whether that index is built off a table, then check if the table is meant to be
persisted, then check if the index is the table's primary index. If all of those checks
are true, we forward the since updates to the persisted table.

We recently introduced a regression where we used the wrong id to check if a table
was persisted - we used the index id instead of the table id. This lead since
updates never being forwarded to persistent tables, which lead to unbounded storage
growth even for persisted tables that only use a finite amount of data.

This commit fixes that regression by looking up "is this table persisted" using the
table id. Further testing to catch this type of compaction error will be done in a
followup commit.