feature-benchmark: Make output sections in CI
This makes the output less than thousands of lines in the buildkite UI.

Each scenario gets its own collapsed section, which both makes it easier to
scan the scenarios and to view the per-scenario report, since it will be at the
end of a section, with only collapsed sections afterwards.

Additionally, if the benchark completes without being OOM- or timeout-killed it
generates a default expanded section with just the report. If there are any
regressions it becomes easy to view the scenario-specific details by expanding
their specific sections.