Remove AWS External IDs from any part of source descriptions
External IDs are part of the runtime environment, not part of the source, so it
makes sense to pass them into AWS clients from the environment instead of from
the DDL description.

This also uses a big scary name to try and make it obvious in future
implementations and code reviews that this data must come from
Materialize-the-operator, not the user of materialized.