Store days separately in Intervals
Previously Intervals stored days, hours, minutes, and seconds all in a
single field. They were all converted to nanoseconds and stored in an
i128. This is different than PostgreSQL, which stores days in a
separate field. This commit changes the internal Interval
representation to match PostgreSQL by storing days separately. It also
changed the i128 nanosecond to a i64 microsecond to match PostgreSQL.

This results in some new behavior around comparing Intervals. Equality
does not convert between months, days, and microseconds. So
INTERVAL '24 hours' does not equal INTERVAL '1 day'. This is a
limitation introduced by dataflow that requires equivalent Datums to
have identical in memory representations.

This commit also makes some small changes to printing and rounding
Intervals to make it closer to the PostgreSQL behavior.

Fixes #10708