correct kafka error message during purification
During purification we create a Kafka consumer to validate that
connection params are correct and that we can reach the Kafka broker. We
potentially use this consumer for fetching things like the start offset
if specified, but that's not its primary purpose.

Today the error message is confusing because it incorrectly references
start offset fetching, even in cases where those params have not been
specified in CREATE SOURCE.

Rewriting the error message to be slightly more generic