postgres-util,pgrepr,sql: redistribute responsibilities for postgres types
Redistribute how PostgreSQL types are handled. Creating a PostgreSQL
type from an OID and typmod is now handled entirely by `pgrepr::Type`.
The `postgres_util::PgColumn` struct now stores a `pgrepr::Type`
directly, rather than a separate `PgScalarType` enum that was relatively
duplicative with `pgrepr::Type`.

The `sql` crate now generates views for a PostgreSQL source by
converting a `pgrepr::Type` to an `UnresolvedDataType` by way of a new
`fmt::Display` implementation on `pgrepr::Type`.

For safety, the modifiers on `ScalarType::{Numeric,VarChar,Char}` now
have newtype wrappers that enforce the bounds at construction. For
clarity, the `scale` parameter of `ScalarType::Numeric` has been renamed
`max_scale` and the `length` parameter of `ScalarType::VarChar` has been
renamed `max_length`, to better reflect that they are maximums, not hard
requirements. This is in contrast to `ScalarType::Char`, where the
length is in fact a hard requirement, not a maximum.