Add support for Array subqueries
Postgres allows array construction through a query, called Array
subqueries. The syntax is `ARRAY(<query>)`. The result is an array
where each element is a row from the query provided. This commit
implements this feature. Currently it has the limitation that the
result of the query itself can't be an array. The reason is that array
subqueries use the array_agg function, which doesn't support nested
arrays.