pgwire: add new send method which filters messages sent to clients
Including the logic on how to map client-side severity levels to regular
pgwire severity levels.

This new `send` method is a wrapper inside StateMachine that always
checks if the message should be sent to a client or not, and if so,
dispatches to the old implementation in FramedConn.

The next commit updates all the callsites to use the new implementation.