persist: correctly downgrade to the empty frontier when the input stream is closed
seal is a tricky operator that uses ProbeHandles to compute the minimum frontier
available for a set of streams that have types not known at compile time. It then
uses one stream as the primary, for which it receives the input frontier as expected,
and then it takes the frontier from all others using the probe handles.

This commit corrects a bug where when we observe the input frontier was empty, we
downgrade to the minimum frontier of the input + all of the additional streams which
is not correct, as we need to instead propagate the empty frontier down.