coord: improve timestamp determination to avoid no complete timestamps
Previously we would require a valid upper to be available and error with
"no complete timestamps" if it was not available. Instead, change the
order in which we determine timestamps:

1. Initialize to the timestamp minimum.
2. Advance to the since.
3. Advance to the upper.

Each "advance" step can possibly do nothing if the current canditate
timestamp is already in advance of the argument.

As a result, we no longer produce the no complete timestamps error. The
previous case where it occurred, when the upper was 0, will leave the
candidate timestamp at the since. If the since is also 0, that is fine,
because we assume the source is correctly declaring 0 to be the first
time at which it will produce results.

Closes #8163
Fixes #3122