repr: (re)introduce `RowPacker`
When using a `Row` in the "row packer" pattern, it was easy to forget to
call `.clear()` in every code path that started a new packing operation.
See, for example, #10767.

To avoid this footgun, force all mutation of `Row` to happen via a new
type called `RowPacker`. Constructing a `RowPacker` *always* clears the
data in the underlying `Row`. Because it is not otherwise possible to
mutate a `Row`, it is no longer possible to forget to call `clear`
before starting a packing operation.

The `RowPacker` type is a reintroduction of a type that used to exist.
However, its new design does not require two allocations to produce a
single row in the worst case. See #5472 and #6276 for some historical
context.