Change process for writing release notes
Change the process for writing release notes so that the notes are
written in the PR description rather than added directly to
release-notes.md. This avoids constant merge conflicts on
release-notes.md.

The release process already entails a complete survey of all PRs to
migrate only the release notes for the PRs included in the release to
the correct version heading, so this doesn't add as much of a burden to
the release process as it might initially seem.