coord: move frontiers creation along side index creation
Previously calls to finalize_dataflow would have the side effect of
creating new index frontiers in the coordinator, assuming an index
would actually be created. However finalize_dataflow was also called
from sequence_peek, which did not call ship_dataflow, and thus we were
tracking an index that didn't exist. Move calls to new_index_frontiers
to ship_dataflows, which is the only thing that creates indexes.

Now that we don't incorrectly attempt to track non-existent indexes,
we can add id validation to allow_compaction in the dataflow client.