kafka: immediately cleanup metadata threads on drop
The metadata thread was previously checking whether the source it was
spanwed for was dropped as the first step in its polling loop.

Since the poll loop by default was 5 minutes it could cause a lot of
idle threads to linger around for a maximum of 5 minutes.

Now the thread parks itself with a timeout instead of sleeping and we
make sure to unpark it when the parent source is dropped. This way the
thread notices that it should exit immediately.

Fixes #10843

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>