*: replace all usage of env_logger with tracing
Continue transitioning our codebase to the `tracing` ecosystem by
replacing all usage of the `env_logger` crate with the equivalent
`tracing_subscriber` invocation.

This results in a slight change in interface: filters are configured via
a `--log-filter` command-line flag rather than the `RUST_LOG` or
`MZ_LOG_FILTER` environment variable. If we need to restore the
`MZ_LOG_FILTER` environment variable, it's trivial to do so with clap
via an `env = "MZ_LOG_FILTER"` annotation on the `log_filter` fields,
but best to avoid environment variables for as long as possible to avoid
spooky action at a distance.