coord,dataflow: remove special cases for table indexes
Remove the special cases for table indexes. Rather than mandating that
each table have a primary index, store the data in a `Vec` outside of
any index. This allows tables to be imported multiple times into various
dataflows, just like other sources.

In the future, the table data will be stored in a persistent location
rather than a `Vec`.

Fix #10720.