dataflow: remove SQL name from SourceDesc
The name of a source is a SQL concept; the dataflow layer should be
entirely oblivious to this name. But abea8395c1 started embedding the
SQL name of the source in dataflow's `SourceDesc` to improve error
messages. This has two defects:

  * The error message includes the wrong name if the source is ever
    renamed.

  * Changing the name of the source causes the source description to
    change from dataflow's perspective, which trips an assertion.

This commit removes the SQL name from the `SourceDesc`. Instead, source
errors report the source instance ID rather than the name, since this ID
is stable even if the name changes. A test is enclosed which would trip
the assertion if not for this fix.

This change, as written, has a negative impact on error messages,
because now the user needs to translate the source instance ID into a
name. We could restore the original behavior by teaching the coordinator
to perform this translation before it sends the error message to the
client.