sql: correct comment about casting string-like types
PostgreSQL allows casting most types to/from varchar and bpchar without
explicit entries in pg_cast. This isn't because PostgreSQL applies
function type promotion rules to casts, as a comment in our codebase
claims, but because PostgreSQL has basically the same special case that
we do. This commit updates the comment accordingly.