Support \d <object> meta command
This commit adds the required catalog tables, columns, and functions
required to support the \d <object> meta command. The majority of these
are not applicable to Materialize and filled with sane defaults. The
rest are implemented in a way that makes sense with Materialize.

The output of \d <object> differs slightly from Postgres. The two
biggest differences are that we include primary key indexes while
PostgreSQL doesn't, and our index syntax is different.

Closes #9595