pg-cdc: fix logic bug with COPY iterator
The implementation of `RawIterator` relied on the implementation of the
protocol parser which when asked to parse a column from the input buffer
returns a `Result<Option<&[u8]>, Error>` where the Option in the Ok case
is used to signal whether the current column contains some bytes or is
null.

The previous implementation of `RawIterator` transposed that result into
an `Option<Result<&[u8], Error>>` changing the semantics in a subtle but
imporant way, because now a `None` value signaled the end of the
iterator.

This resulted with the ingestion pipeline early returning when consuming
the COPY data and omitting all column data after encountering one with a
NULL.

This PR fixes the problem by making the iterator return
`Option<Result<Option<&[u8], Error>>` where the outer option signals the
end of the Iterator and the inner option singal the presence or not of a
NULL in the column.

Fixes #10981

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>