repr,pgrepr: add PostgreSQL "char" type
The "char" type (note quotes) is a one-byte type that is used in various
pg_catalog tables. It has a distinct binary representation from strings,
so our approach of using `char(1)` as the representation was causing
problems with drivers that use binary encoding, like rust-postgres.

This commit introduces a "char" type to Materialize that exactly matches
the PostgreSQL semantics. We call it `ScalarType::Byte` internally,
since `ScalarType::Char` is already in use for what PostgreSQL calls
"bpchar".