coord: declosureify Session::get_transaction_timestamp
Refactor `Session::get_transaction_timestamp` so that the method just
returns the transaction timestamp, rather than potentially *determining*
the timestamp with caller-supplied logic. The caller is now responsible
for setting the transaction timestamp if it comes back `None`.

There is only one caller, so this commit is straightforward code
movement. The motivation is easing some changes in #11019, which need to
make some `async` calls during transaction timestamp determination.