testdrive: don't kill materialized after a successful test run
While debugging, sometimes you want to observe the state of materialized
even after a successful test run. Testdrive scripts in this directory
can be run repeatedly against the same `materialized`, so no need to try
to clean things up; the start of the next test run will do that if
necessary.