dataflow-types: split out immutable controllers
Split the `StorageController` and `ComputeController` types into mutable
and immutable variants. This is useful in several `Coordinator` code
paths where we want to call immutable methods on the controllers but it
is distasteful (or impossible) to get the required mutable references to
construct the controllers.

An immutable controller can be constructed out of a mutable controller
via the `as_ref()` method. (You might hope to use `Deref`, but that's
not possible due to the lifetimes in the trait.)