dataflow-types: beef up IndexDesc and use it more
Add two additional fields to the `IndexDesc` that are often stashed
alongside it: the ID of the index itself and the type of the relation on
which the index is built. The only cost is that the `index_imports`
field now stores index IDs twice, but that seems like a small price to
pay.