Add SQL parsing and plumbing for Loki source creation
This adds the new keywords to the SQL parser for the Loki source, and
connects everything up so that a Loki source is created by the CREATE
SOURCE query.

This is almost all copy+pasted from the PubNub source (or just driven by
enum exhausiveness-checking). Some of the values may need changing soon
as the source develops.

The CREATE SOURCE command looks like this:

    CREATE SOURCE src_name
    FROM LOKI
    ADDRESS 'http://loki'
    QUERY '{job="varlogs"}'
    FORMAT TEXT;

Note that the only two source parameters here are 'address' and 'query',
which differs from the three commands in the constructor of the
`LokiSourceReader`; these may need changing at a later date.

The Loki source also only supports FORMAT TEXT for now, on the
assumption that we'll always be fetching JSON from Loki.