dataflow: Only send frontier uppers for non-empty lists
Currently, we as send many `ComputeResponse::FrontierUppers` to the
coordinator as there are Timely workers. This is not required as the
partitioned client absorbs updates from all its Timely workers anly only
releases combined frontier information.

With this change, we only pass the command to the coordinator if the list
of frontiers is non-empty.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>