[kafka sink] Retry message send failures (#10896)
Attempt 2 to replace #10853 

There are two types of failures we can see when sending a message to rdkafka:
1. The `send` call itself.  This enqueues the message on a buffer to be sent later
2. The actual sending itself (which happens on a background thread)

We were handling (1) but not (2).  My current approach is to expand the responsibility of `KafkaSinkState::flush` to make sure that all messages are successfully sent.  This involves two new structures:
1. Retry queue: where the messages to be resent go after they are rejected
2. Tracking of outstanding message sends
We need both of these because the `ProducerContext::delivery` callback is made asynchronously after the message is successfully or unsuccessfully sent.  We need to be able to know when it's no longer possible to get more errors.

### Motivation
Fixes #10755 which alerted me to the fact that we didn't properly handle errors during the actual sending of a message.

### Checklist
- [x] This PR has adequate test coverage / QA involvement has been duly considered.
- [ ] This PR adds a release note for any [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/user/content/release-notes.md#what-changes-require-a-release-note).