materialized: refactor panic handling setup
@guswynn noticed in #11107 that our custom panic handler will sometimes
fail to report a panic. The issue is that the panic hook calls
`tracing::error!` but is installed before tracing is configured; that
means that a panic before tracing is configured will shout into the
void.

Refactor the startup sequence to avoid that. Specifically, configure
tracing *before* installing the custom panic hook.

This commit is pure code movement, except for a few new comments that
explain what is going on.