Modify planner to work with augmented statements
Previously all planner functions would take in raw statements and
resolve names on an ad hoc basis as needed. This commit updates this
behavior so that the planner resolves all names as it's first step and
the planner functions take augmented statements. There are two
exceptions to this:
  * SCL planner functions still work with raw statements
  * SHOW and DROP statements use augmented statements, but the
  augmented statements contain unresolved names.
These exceptions exist not for any logical reason but to try and break
up the work into multiple smaller commits.

These changes gets us closer to removing some of the limitations on our
ALTER statements, by starting to allow us to represent dependencies
with IDs instead of names.

Works towards resolving #5302