coord: improve table timestamp generation
For all writes that are ready to commit (which does not include deferred
writes waiting on `write_lock`), add them to a queue and immediately try
to process it. If a timestamp in advance of previous timestamps cannot
be produced, try again later. This allows us to stop using `ts+1` logic
in the timestamp oracle to manually force timestamps forward by 1ms even
if 1ms of time has not occurred (or the system clock has retreated).

Plumb now through more of the mz infra so we are able to inject clock
failures during testing.

Closes #10173
Fixes #8424