pgcopy: invert dependency with sql
Make sql depend on pgcopy, rather than vice-versa. This prevents changes
to sql requiring that the dataflow crate be recompiled, which is slow.
(dataflow depends on pgcopy).

Doing it this way is also keeps the responsbility for translating a SQL
AST into a more structured type *in* the sql crate. The previous
CopyParams struct exported by the sql crate was basically just the AST
for the COPY statement, and the responsibility of validating that AST
was kicked downstream (i.e., to pgcopy). Better for pgcopy to assert its
needs (via the properly-structured CopyFormatParams type) and let the
sql crate do the hard work of producing that struct.