testdrive: Apply timeout to the individual SQL statements
Previously, only Retry was used to enforce timeouts, which meant
that if a single execution of a SELECT statement hung, testdrive will
hang permanently as the control will never get back to Retry so that
the timeout can be enforced.

So, use tokio::time::timeout to time out individual SQL statements as
tokio-postgres only supports a connection timeout.