sql: remove catalog from purification
This brings the signature of the `purify` function back in line with its
documentation. Purification is meant to be an operation that does not
depend on the catalog at all; we corrupted this to support `CREATE VIEWS
FROM SOURCE` in a hurry. Fortunately the refactor in #11083 allows us to
restore `purify` to its original ... purity.