coord,sql: install linearized initial timestamp bindings
When creating a source, as part of purification, fetch the latest known
offset, and assign an initial timestamp for that offset. This ensures
that the data that is present in the source at the time of its creation
appears in Materialize in one batch at the first queryable timestamp.

Fix #11048 and #11072.

Co-authored-by: Frank McSherry <fmcsherry@me.com>