coord: disable SELECTing from non-table unmaterialized sources
There are known semantic glitches with unmaterialized sources that can
cause them to present partial (or no) data when SELECTed. So, restore
the behavior that was present before #11037 by rejecting SELECTs that
reference unmaterialized sources.

The slight tweak in the new behavior is that tables are exempted from
this requirement. Tables are linearized by the coordinator, and so
SELECTing from an unmaterialized table does not have the same defect
that SELECTing from other types of unmaterialized sources do.

This makes #11072 and #11232 no longer release blockers. The underlying
issues are still visible via `AS OF now()`, but we're willing to live
with that now, to give us time to solve the semantic defects properly
(see #11048).