scratch: Print the region that instances are in
Because of AWS UX, knowing the exact region is pretty important -- trying to
operate in the wrong region will just turn up "not found" or even worse
responses (e.g. for aws ssm it will say that the ssm client on the vm did
not start, not that the vm does not exist).