feature-benchmark: Add scenario for measuring startup performance
New scenarios are hereby added that checks the elapsed duration between
the return of the `docker-compose up` command that is used to start Mz
and the time Mz begins servicing queries.

The `StartupEmpty` scenario creates no database objects and considers the
database to be up as soon as SELECT 1 returns.

The `StartupLoaded` scenario creates 100 materialized sources, 100 sinks ,
100 tables and 100 3-table materialized views. It considers Mz to be up
once the sources and views start returning at least one row. Tables
are expected to return 0 or 1 row, as this depends on whether persistence
is in effect.

In addition:
- it is now possible to have mzcompose operations as part of the workload
  being benchmarked, not just in the init() or before() phases as it was previously.
- the feature-benchmark has been extended to operate on timestamps
rather than measurements. This way, one part of the workload can provide
the start timestamp and another part can provide the end timestamp.
The framework will then calculate the elapsed time.
- the measurement_source.py no longer supports an Iterator interface
- various dead code has been removed