dataflow: swap Client::recv for Stream
The `Client::recv` method is more idiomatically a `Stream`
implementation in Rust async/await. Modeling it as such makes it easier
to compose the `Client` with existing stream combinators.  It also
avoids the pattern in `SelectStream` which incorrectly constructs a new
future on every call to `poll_next`. This defect was spotted by
@petrosagg in #11223; the code happens to work with the current stack of
futures/streams, but might break at any point in the future.

This starts to drive towards the goal laid out in #11078. We don't
actually get separate `Stream` and `Sink` types in this PR, but we at
least get the `Stream` implementation.

Accounting for this refactor in `coordtest` was too difficult to be
worth it, so I opted instead to wholesale remove `coordtest`, as the
comment within `coordtest` offers.