materialized: absorb coordd
Teach `materialized` to absorb all the responsibilities of `coordd`, and
then delete `coordd`. This is sort of like #11132, except less stressful
from a security point of view, as this commit is careful not to open any
new ports in `materialized`.

`materialized` always hosts a single compute runtime to host virtual
clusters. All vitual clusters are created in this in-process runtime.

By default, `materialized` also hosts a storage runtime in the same
timely cluster as the compute runtime.  Optionally, the new
storage command-line flags

  * --storage-workers
  * --storage-compute-addr
  * --storage-controller-addr

can be passed to specify that `materialized` should not start a storage
timely cluster, but instead connect to the storage process.

`CREATE CLUSTER` also learns to take a new `REMOTE` option, so that it
is possible to test connecting to a cluster that is hosted in a
different process. This option will never be exposed to users, but it is
useful for testing now.

The `LegacyMultiplex` mode is removed from `dataflowd`, as it was no
longer (easily) possible to use it with `materialized`, as
`materialized` would try to connect twice: once via the storage client
and once for the compute client. The philosophy is that either we run in
single-process mode or fully-distributed mode; there's not much need to
support a semi-distributed mode. The cluster tests are updated to only
exercise the fully-distriburted mode.

A large part of this commit is refactoring the `Controller`/`Client`
stack so that the `Controller` manages separate storage clients and
compute clients, with one compute client per compute instance.
Previously, the `SplitClient` type would snoop commands as they flowed
past and manufacture new clients on the fly. The new approach makes the
structure of the `Controller` more logical, at the cost of some
complicated async code that multiplexes commands and demultiplexes
responses. This code will vanish, though, if we eventually get rid of
virtual clusters.