materialized: disable persistent system tables test by default
As things are rapidly changing in the transition to platform, it was
assumed that we'd accidentally break this test along the way. We could
fix it, but the value we'd get from keeping the test working during the
transition (some value) is not worth the work (dunno exactly, but it's
too much work) compared to disabling the test and re-enabling it once
things have settled down.

Today is that day. Philip found #11332, which is sufficiently worrisome
that it's no longer worth keeping this test on in the short-term. This
changes the default for the flag to disable the test.