persist: Introduce candidate API for consensus.
This commit is not for merging and only to start discussion. This implements a
candidate API for Consensus, which I think, we want to represent a
register that stores:

(SeqNo, Vec<u8>) where a user can modify the data to (new_seqno, new_data) and the operation
will succeed iff the current seqno == new_seqno - 1 (aka cas to guarantee total order on
updates across multiple writers), but the sequence of changes to the data across sequence numbers
is also readable, and truncatable.

This API is itself built on top of one of two potential APIs that are less expressive

CasKeyVal - which describes a key value store with a CAS/conditional put operation, but without
the ability to see the history of updates to any one key (e.g. DynamoDB's PutItem API for a real world example)

AtomicQueue - which descibes a queue that allows for atomic appends, and scans / truncations from an
index (aka offset) within the queue but does not have a builtin CAS operation (e.g. Kafka for a real world example)

I believe, according to [1] the two lower level APIs are equivalently expressive and either of them can be
used to implement Consensus as defined in this PR.

[1]: https://cs.brown.edu/~mph/Herlihy91/p124-herlihy.pdf