dataflow-types: make TcpClient::recv cancellation safe
Previously, cancelling a `TcpClient::recv` future could result in
aborting an otherwise successful connection attempt, resulting in a
"wasted" connection that caused the receiving dataflowd process to
terminate. Rewrite `TcpClient::recv` in a state machine style that is
cancellation state.

This makes the cluster tests much more reliable. There are still some
outstanding issues with `test/cluster`, so I'm leaving the issue open
and the test marked as soft_fail.

In particular, there are probably other clients in the stack
(`ActiveReplication`?) that are not cancellation safe. I'm not sure that
making `Client::recv` cancellation safe is the right long-term plan, but
it's the shortest path I could envision to making the demo this week
somewhat reliable.

Touches #11390.