dataflow-types: teach PartitionedClient to absorb errors
The PartitionedClient is meant to abstract away the details of
communicating with a multi-process timely cluster. Previously the
multi-process nature leaked out when errors occurred: one error was
returned for each process.

This commit teaches PartitionedClient to instead suppress all but the
Nth error, where N is the number of parts. This ensures that the caller
observes each error only once.

Notably, this PR, along with the already merged #11405, fixes #11390.
The issue was that the `ActiveReplication` client would observe *two*
errors on initial connection, for the reason mentioned above,  and
attempt to hydrate the replica twice, which is not yet safe. #11103 will
one day make hydrating the same replica twice safe, but we're not there
yet.