persist: initial Consensus trait definition + in-mem implementation
This commit introduces the Consensus trait which is defined as an abstraction over
a persisted, arbitrarily sized array of binary data + a version number. The data
can be conditionally updated based on the version number.

We will use this API to get a total order on the evolution of persist-internal
metadata cross multiple processes and machines.

This commit also provides a bare bones in-memory implementation and unit tests.