storage: introduce StorageController trait
allow other compontents to abstract over an implementation of a
StorageController. This allows controller implementations that for one
reason or another have extra constraints on their types and we don't
necessarily want those constraints to contaminate the components that
generically interact with a StorageController.

This came up as part of #11223 where the StorageController
implementation became specific over i64/u64 timestamps and those
constraints contaminated COMPUTE.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>