coord, sql: partial support for DROP SECRET and ALTER SECRET .. RENAME TO .. (#11362)
Implement DROP SECRET command. The implementation is partial and does not call into the secrets backend. It cleans up the catalog only.

Implement ALTER SECRET .. RENAME TO ... Given that rename does not change the secret on disk and only changes SQL metadata, this functionality, in a sense, is complete.

ALTER SECRET .. AS .. remains unimplemented.