Modify planner/catalog to use database/schema ids (#11313)
Previously databases and schemas were primarily referred to by their
string names. This commit modifies the code to primarily refer to
schemas and databases by their IDs instead of their names.

These changes gets us closer to removing some of the limitations on our
ALTER statements, by starting to allow us to represent dependencies
with IDs instead of names.

Additionally this commit happens to clarify some errors when the active
database was deleted. Previously the active database for a session was
stored as a string, and it was never verified that it existed. When
creating objects, if a database is not specified, the active database
is used by default. Therefore if you tried to create an object without
a database specified and the active database was deleted, a cryptic
error would be printed about not being able to find the database. Now
the active database is stored as an Option<DatabaseId>, so the error
message can indicate that there's no active database.

Additionally SHOW CREATE VIEW for builtin views now properly display
the correct system schema, instead of incorrectly displaying the
default database and schema.

Works towards resolving #5302

Fixes #11343