sql: Dynamically allocate system global IDs
Previously the global IDs for built-in system objects were statically
compiled into the Materialize binary. Dataflow operates on the
assumption that a single global ID only ever refers to a single object.

In the single binary world this is fine, because when built-in object's
schema changes, the new schema can be picked up by everyone on a
restart. In the platform world this is a problem because the ADAPTER
can restart independently from STORAGE and COMPUTE, and the ADAPTER
needs a way to tell STORAGE and COMPUTE about the new schema and assign
it a new global ID. Assigning a built-on object a new global ID is very
difficult if that global ID is compiled into the binary.

This commit updates the ADAPTER to assign global IDs to built-in
objects dynamically on start-up. This will make it easier to change a
built-in object's ID when it's schema changes.

Actual schema migrations and updating a built-in's global ID are not
yet implemented and left to a later commit.

Built-in types are still statically assigned global IDs, because they
are slightly more complex than other objects due to the fact that they
have references (possibly circular) to other built-in types. This
doesn't create a huge problem because types have no schema and
therefore will never have schema migrations. We leave dynamic IDs for
built-in types as future work.

Works towards resolving #11435