coord,dataflow,sql: remove virtual clusters
This commit removes the concept of virtual clusters. They were useful as
a way to plumb compute instance IDs throughout the codebase before we
had remote and managed clusters. But now that we have these real types
of clusters, virtual clusters are getting in the way.

Specifically, virtual clusters meant that a single timely instance could
host multiple compute instances, which meant that large swaths of the
system needed to be multi-cluster aware. By removing virtual clusters,
only the coordinator and controller need to be multi-cluster aware, and
the downstream components only need to be aware of a single cluster.

To preserve the behavior of the binary, the `default` cluster is now
special cased as a "local" cluster. It always exists, is always
connected to the local timely instance, and cannot be dropped. It is not
possible to create another local cluster: you only get the one.