test/metrics: merge into test/testdrive
On reflection, having a separate mzcompose composition for the one
metrics testdrive script doesn't seem worth it. The composition itself
is a duplicate of test/testdrive, and the extra CI job adds 30-60s of
overhead.

If we need to test metrics for e.g. PostgreSQL in the future, we can add
that script to test/pg-cdc, so that we can reuse the existing CI job and
composition.