dataflow: split storage and compute clients in server
This commit changes the dataflow server to expose separate storage and
compute clients, rather than multiplexing storage and compute commands
and responses on the same client. This is a step towards the goal of
separating storage and compute. In particular:

  * The logic for partitioning commands and responses across multiple workers
    is teed up to be layer local, by way of separate `PartitionedComputeState`
    and `PartitionedStorageState` types.

  * The compute-specific reconciliation logic no longer needs to panic
    if it sees a storage command or response; the type system can
    provide this safety.

  * The `ComputeWrapperClient` and `StorageWrapperClient` are similarly
    made unnecessary, as the dataflow server exposes `ComputeClient`s
    and `StorageClient`s from the get go.

  * The sketchy async code to demultiplex the combined client (`split_client`
    and `ChannelClient`) is entirely deleted.

The server metrics were frustrating this change, as they assumed the
presence of a single dataflow server command queue; this is no longer
the case with the separate storage and compute command queue. These
metrics weren't that useful--we've never really cared about the total
number of commands processed by type, for example--so rather than
attempting to fit these metrics into the new model, this commit just
deletes them. In platform, these metrics won't be user-facing anyway; we
can add back what we need if it proves necessary for debugging.