dataflow: restore single queue for dataflow commands
The last commit had a glitch, in which storage and compute commands
could be interleaved in different orders on different workers, which
could cause timely to panic because dataflows were built in different
orders on different workers.

Restore determinism by forcing all storage and compute commands into a
single channel which enforces a (arbitrary) total order that is
consistent across workers.

This backpedals a bit on the separation of storage and compute, but only
minorly. The layers above the dataflow server still observe fully
separated storage and compute clients.