deps: unfork rust-prometheus
Get back on the official release of rust-prometheus. Move the one
still relevant patch from our fork--unsigned gauges--into ore::metrics.
IntCounters are now unsigned by default upstream, so we can replace any
usage of UIntCounter with IntCounter.

This commit also removes the process-collector crate. This crate was a
copy/paste of the process collector in rust-prometheus, with the
addition of swap metrics. Just wholesale remove the custom process
collector, since swap metrics are no longer nearly as relevant in
Materialize Platform. If we need to start tracking swap again, we can do
so in a way that does not require copy/pasting so much code.