Validate data dir in database copy (#11582)
Testdrive has an option, --validate-data-dir, that validates the
contents of the SQLite database. It does this by opening a new catalog
using the database directory and comparing the contents against the in
memory catalog.

Opening a new catalog using the same SQLite database that Materialize
is actively using can cause some conflicting database operations.
Therefore this commit updates Testdrive to open a new catalog using a
copy of the SQLite database to avoid conflicts.