persist: make API structs Send + Sync
This makes them usable in async tasks, where Futures need to be Send +
Sync.

We add the constraints to where traits are used as opposed to adding
them on the traits directly because the constraints come from the API
and are not needed on the BlobMulti and Consensus traits in general.