stash: lighten the load on SQLite
This patch simplifies the schema of the SQLite backend of stash by
foregoing consolidation on insert which involved a DELETE statement and
a unique index.

The external API remains unchagned since we were consolidating the
result of the stash implementation anyway.

Consolidation logic has been adapted too to account for the lack of
unique index which was driving the logic before.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>