coord: Detect schema changes for builtin objects
This commit allows us to detect when the schema of a built-in object
changes. Schema here refers to the column names and types of a built-in
object, not the namespace within a database. This brings us one step
closer to being able to perform schema migrations when a built-in
object changes schemas.

This commit doesn't actually perform any schema migrations, it just
adds the plumbing necessary to detect when they're necessary. Schema
migrations are left for a later commit.

Works towards resolving #11435