sql: Clean up coord bootstrapping
Previously opening the catalog would return a list of log variant,
global ID pairs. The coordinator would use these pairs to help bootstrap
certain catalog tables.

This was a bit messy, and all the info that was passed from the catalog
to the coordinator could be looked up on an ad hoc basis as need. So
this commit removes that list and updates the coordinator to look up
log info in the catalog.