coord: add view definition to `{mz,pg}_views`
This commit adds a `definition` column to both `mz_views` and
`pg_views`. This column contains the SQL SELECT statement that was used
to create the respective view.

Note that the `definition` column does *not* contain the whole CREATE
VIEW statement. This is consistent with the PostgreSQL implementation
(https://www.postgresql.org/docs/8.0/view-pg-views.html).

We introduce this new column to enable the implementation of the
`pg_get_viewdef` function (#11591).