sql: Dynamically allocate builtin type global IDs (#11643)
commit 9859e61723238dedddf5c367acefda74b4a46722 switched from static
global IDs for all built-in objects to dynamically allocated global IDs
for all built-in objects, except for built-in types. Built-in types
reference other built-in types, sometimes with circular references. The
references use global IDs, which made it difficult to dynamically
assign global IDs to built-in types while correctly maintaining those
references.

This commit adds the functionality to dynamically assign global IDs to
built-in types, so that all built-in objects have dynamically assigned
global IDs.

Works towards resolving #11435