orchestrator-kubernetes: don't crash if pod doesn't exist
The code added in #11635 to force Kubernetes to replace stateful set
pods in all cases was not quite careful enough in handling the case
where the pod did not already exist. This would cause `materialized` to
crash 50% of the time when creating a new cluster, depending on how
quickly Kubernetes created the pod.