sql: Clean up catalog state on cluster drop
Previously when a cluster was dropped we forget to remove it's entries
from the SQLite catalog table "compute_introspection_source_indexes".
This made it impossible to create a cluster with the same name as a
previously dropped cluster, if they both had introspection sources
enabled.

This commit fixes that issue by cleaning up entries when a cluster is
dropped.