dataflow-types: reconfigure Response
The `Response` type is currently an enum that contained all storage
responses and all compute responses. But some of these responses are
meant to be handled entirely by the controller; i.e., not all of them
are meant to be forwarded on to the coordinator for additional
processing.

This commit makes the protocol explicit in the type system. The
controller now returns responses of type `ControllerResponse`, which
contains a subset of the possible storage and compute responses. In the
future, `ControllerResponse` can diverge even further from
`StorageResponse` and `ComputeResponse`, e.g., when a response is
generated entirely in the controller and does not require communication
with a storage or compute instance.