platform: Make linger configurable
A previous commit removed the ability to set the "--linger" option for
computed and storaged. This commit adds a flag to materialized called
"--orchestrator_linger" which will set the "linger" flag on computed
storaged when they are started.

If "--orchestrator_linger" is not set, then "--linger" will have
different defaults depending on the specific type of orchestration.
Process based orchestration will default to "--linger" being false to
make local development cleanup easier. Kubernetes based orchestration
will default "--linger" to true, so that COMPUTE and STORAGE nodes are
resilient to ADAPTER failures.

All tests set "--linger" to false for now, though we will likely want
to add tests in the future that set "--linger" to true.