bin/cluster-dev: set image pull policy to never
When developing using bin/cluster-dev, the
materialized/storaged/computed images will not be available on Docker
Hub; they'll have been pushed to the local Docker daemon directly. This
is incompatible with the default image pull policy of "always" that we
want to use in prodution. So, expose the image pull policy as a
command-line flag, and set it to "never" when running via
bin/cluster-dev.