ci/test: allow test jobs to run on builder agents
Our CI agents are split into "builder" and "non-builder" agents. The
idea is that the builder agents are a limited set of beefier that
maintain warm target directories, build the requisite set of Docker
images, and then farm the actual running of tests out to other agents.

But we're frequently not saturating our pool of builder agents with
jobs. This commit adjusts the test pipeline queue targeting so that test
jobs can run on builder *or* non-builder agents. It adjusts priorities
so that build jobs still have precedence on builder agents.