bin/cluster-dev: force docker-env output for bash
It appears that `minikube docker-env` selects its output format based
on the $SHELL env variable. $SHELL does not specify the current shell,
but the default shell. So on a system that uses a POSIX non-compliant
shell, like fish, `minikube docker-env` outputs a format bash does not
understand. To mitigate this, set the target shell to bash explicitly.