dataflow-types: maybe deflake Kafka resumption tests
PR #11846 was the first PR to fail the Kafka resumption tests. On a
close look, it was accidentally behavior changing, in that a `None`
response from the storage instance caused further messages from the
storage instance to be ignored.

This commit restores the previous behavior in which a `None` response
from the storage layer is entirely ignored, and further `Some` responses
that may occur are processed. This seems in violation of the
documentation on how `recv` works (`None` indicates permanent shutdown),
but worth a shot.