dbt-materialize: stop using Avro OCF sources/sinks
Avro OCF sources and sinks are being removed since they aren't usable in
Materialize Cloud (#11875). This commit changes the dbt test suite to
use Kafka sources and sinks instead.

We lose a little test coverage here because we can no longer verify that
the source actually *works*, but that's okay; we've got plenty of other
tests in Materialize for that. We'll get the test coverage back when
Platform fully supports EOS sinks, and then we can test that a sink that
emits static data can be reingested with the same data, all in dbt.

Touches #11875.