tcp_boundary: drop source correctly
Instead of waiting to schedule ok and err streams to drop the source, only
require a single of the streams to be terminate to terminate both. This
prevents a situation where the ok stream is terminated, but the err stream
will not pick up the termination because it won't be scheduled anymore.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>