testdrive: port file delimiter test to s3
File sources are slated for removal (see #11875). Port a test that uses
file sources to test how the source ingestion pipeline handles trailing
newlines to use S3 sources instead, as we still want to test this part
of the source ingestion pipeline.

This requires changes to testdrive to make the behavior of
`s3-put-object` match the behavior of `file-append`.