sql: Remove IF NOT EXISTS from CREATE CLUSTER statement
CREATE IF NOT EXISTS (CINE) features tend to be a debated feature in
SQL. Searching "CINE" on the postgres hackers mailing lists reveals
some past discussions about this feature. In general after executing
a CINE statement, the state of the system can be ambigious if the
details of the create statement don't match the details of the existing
object.

This commit removes the IF NOT EXISTS feature from CREATE CLUSTER since
we don't currently have a use case for it.