stash: don't wrap sqlite conn in Arc Mutex
This allowed the sqlite impl to be Clone'd, but that's not a thing that
we even do. We've internally discussed multiple times that users of stash
are responsible for ensuring only one thing is happening on a stash at
once, so disallowing cloning seems perfectly fine. Also, the upcoming
postgres with an in memory fence counter means we even more don't want
to need cloning, so this seems like a good direction to go.