materialized: Prevent reallocating ports in use
The process orchestrator uses an IdAllocator to allocate ports to
every process. Previously, when the processes orchestrator connected to
existing already running processes, it had no way to inform the
IdAllocator of the ports used by the running processes. That meant that
the IdAllocator may mistakenly allocate ports used by those existing
processes.

This commit adds the functionality to the IdAllocator to mark a port as
already allocated. When the process orchestrator connects to an
existing process, it marks all ports used by that process as allocated.