adapter: reconcile storage commands
When the ADAPTER starts up, it attempts to re-create all sources and
populate all built-in catalog tables. If STORAGE has also just started
up, then this works fine. However if STORAGE was already running and
installed all the sources from a previous ADAPTER, STORAGE will crash
because it already has all of the sources created.

This commit attempts to resolve the above issue by adding a
reconciliation layer in front of the storage instance. The
reconciliation layer will absorb previously seen commands.

The ADAPTER will always clear the contents of all built-in catalog
tables and re-populate them from scratch every time it restarts.

This commit does not attempt to reconcile the RenderSources command
because that command is only sent between STORAGE and COMPUTE. So it is
saved for a later commit.

Resolves #12096