persist: rename Client to PersistClient
There are going to be lots of Clients in the codebase soon. Instead of
requiring everywhere to rename it on import, just name it unambiguously
in the first place.

For the same reason rename Location to PersistLocation.

Leave ShardId alone because persist is the only thing with a shard
concept. Also leave ReadHandle, WriteHandle, Listen, SnapshotShard, and
ShapshotIter alone because their use outside persist should be localized
enough that it shouldn't be a problem and renaming them would really
clutter the persist codebase.