coord: remove support for NOW() in AS OF
`AS OF now()` has historically been used and recommended to get around
some timestamp selection problems we have had. At this time we're not
aware of any existing, valid uses for this syntax that solve a specific
problem. Additionally, we are aware of problems this syntax does cause
due to mismatches between what `now()` evaluates to and how that does
not always match the assumptions that previously ingested data has been
assigned timestamps less than that.