persist: rework retries
This adds indefinite retries (and removes the ExternalError return type)
around every persist call with three exceptions. A caller is free to
drop the persist future to cancel it and we'll (eventually) make sure to
propagate that cancellation to any work happening on it's behalf. As a
result, the tokio::time::timeout Future is now the recommend way to use
timeouts with persist.

Pushing retries into persist wherever we can provides mild performance
benefits but more critically, allows us to retry some errors that
couldn't be retried at a higher level. In particular, during a
WriteHandle::compare_and_append call, it will always be safe to retry
when writing to Blob (even indeterminate errors) but not to Consensus
(in the case of indeterminate errors). The distinction of where an error
came from will have been lost by the time it gets to the persist caller.
On the other hand, there are a few select places where it's not
safe/sensible for persist to retry, but the caller has context on
whether it's safe for not. In these few places, the ExternalError is
still bubbled up.

The three exceptions:
- Location::open doesn't know if the external durability stores are
  correctly configured, permissions are set, etc. Once we introduce a
  determinate/indeterminate split for ExternalError, Location::open
  could also retry indeterminate errors and only send back determinate
  ones.
- WriteHandle::append will be done in a followup commit.
- WriteHandle::compare_and_append is not always safe to retry depending
  on the usage pattern. We should be able to structure timestamp
  binding, source, and sink code so it's always safe to retry, but SQL
  txns will have to pass the error back to the user (or risk double
  committing the txn). Once we introduce a determinate/indeterminate
  split for ExternalError, WriteHandle::compare_and_append could also
  retry determinate errors and only send back indeterminate ones (the
  opposite of Location::open).

Left for followup work:
- A bit more polish on idempotence and testing for idempotence.
- Testing cancellation on drop and propagation of that.
- Automatic internal lease renewals (once we have leases).