persist: make state size proportional to non-empty batches
Sneaky optimization! It was previously proportional to the number of
batches (empty or non-empty). If there are no keys in a batch, we can
simply extend the description of the last one (assuming there is one).
This will be an absurdly common case in actual usage because empty
batches are used to communicate progress and most things will be
low-traffic.

To make things easier to reason about, we avoid mutating any batch that
actually has data, and only apply the optimization if the most recent
batch _also_ has no keys. We call these batches with no keys "padding
batches".

Extension of work in #12225. After this final fix, a test writing 64_800
batches, mostly empty but with 1KiB of data every 1_000, into file blob
and sqlite consensus used 260KB (66 files, exactly one per batch) in
blob and 27KB in consensus (down from 7MB after 12225 and 80GB before
this work). The runtime of the test also went from ~800s after the
previous PR down to ~30s (unmeasured before this work).