expr,test: Add expressions for defining and activating failpoints
Create a pair of functions:

* mz_internal.mz_set_failpoints(str)

Configures the failpoint string, similar to SET failpoints = ...

* mz_internal.failpoint(str)

Creates a failpoint with the specified name that will be hit
when the failpoint() function is evaluated.

If the functions are fed constant arguments, they will be evaluated
by the coordinator and will apply to the process in which the coordinator
is running.

If they are used in a context where a dataflow would be constructed,
such as a materialized view, and given a non-constant argument, they
will be evaluated by the dataflow and will apply to the computed process
that happens to evaluate them.

The functions return the string they were passed as an argument.