doc/developer: remove legacy release process documentation
The release process for v0.26 LTS is now considered legacy. We'll need
to develop a new release process for Materialize Platform.

This commit removes the documentation for this legacy release process,
so that it doesn't clutter up the `main` branch. It still exists on the
v0.26 branch for when we need to cut a new release of v0.26 LTS.