storage: replace MessagePayload with Option<Vec<u8>>
Traditionally the `MessagePayload` enum was used to represent the output
of non-delimited sources, i.e sources that output a stream of bytes.

The pattern of stream of bytes is already established in Rust by either
the family of `Read` traits, that fill up a provided buffer and return
the number of bytes written, or the `Iterator`/`Stream` traits that
return an `Option<Self::Item>` where the `Item` would be a `Vec<u8>` in
this case and `None` would represent the end of the stream.

Our own enum doesn't provide additional semantics than the ones conveyed
by the standard practice so this PR changes non-delimited sources to
produce `Option<Vec<u8>>` messages.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>