persist-client: follow Iterator convention in SnapshotIter
`SnapshotIter`'s API relied on the caller to check if the vector is
empty to signal the termination of the iteration which lead to
non-idiomatic code at the call-site.

By returning an `Option<Item>` callers can use the `while let Some(..)`
pattern to iterate over the snapshot with simpler control flow.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>