[backport] materialized: Allow setting a password prefix
In order to participate in github's secret scanning program, we need
to be able to identify potential platform passwords. This is easiest
if they come with a fairly unique prefix, and that requires that the
materialized protocol be able to detect and enforce that prefix.

This change adds a configurable prefix for frontegg auth and removes
the prefix (erroring if it's not present).