coord: Fix panic on concurrent DDL (#12305)
Read transactions have read holds on certain GlobalIds.
These read holds are maintained by to Coordinator in the txn_reads
struct field. The Coordinator also maintains read capabilities for each
GlobalId in a struct field called read_capability. When a table or
source is dropped, the GlobalId is removed from the read_capability
field, but the read holds by any outstanding transactions are not
removed. This means that when the transaction goes to commit, it will
try and remove a read hold from a non-existent GlobalId and panic.

This commit updates the commit logic to only remove read holds from
GlobalIds that still exist.

If the read transaction tried to read from the GlobalId after it was
deleted, then it would have already ABORTed with an unknown item error.
If the read transaction didn't try to read from the GlobalId after it
was deleted, then it will COMMIT successfully. Neither of these
scenarios violate any consistency guarantees.

Fixes #12283