coord: Make all write durable
Previously consecutive writes in the coordinator were all done at the
same timestamp and saved in an in-memory WAL. These write would only be
persisted to STORAGE after 1 second or after a read. If the ADAPTER
crashed after the write, but before 1 second or a read, then the writes
would be lost forever.

This commit updates the ADAPTER code so that every write increases the
global timestamp and calls append with the write.

In the future if we were to add a durable ADAPTER side WAL, then we
wouldn't need to increase the timestamp or call append for every write.
Writes would only need to be made durable in the WAL, and could be
batched together in a single append call.

Fixes #12287