Create compute cluster pods in parallel
The default StatefulSet pod management policy is to bring them up in
sequence: However, if a 2xlarge cluster requires 12 pods and needs to
autoscale up for every one of them, this is unnecessarily slow.

Instead, let's set the policy to "parallel" (materialized can manage),
and tell the k8s management layer immediately how many pods to scale to.