materialized: add slim Docker image variant
Add a new Docker image variant called `materialized-slim`. This image
contains *only* `materialized`, and does not bundle `storaged` or
`computed`.

The idea is that this image will be used in Materialize Cloud where
`storaged` and `computed` are spawned via Kubernetes in separate pods.
The local process orchestrator will not work with `materialized-slim`,
as it will not be able to find `storaged` or `computed`.

The `materialized` image will remain as an all-in-one option for use in
developer environments and CI pipelines.