persist: split ExternalError into Determinate and Indeterminate variants
When a BlobMulti or Consensus impl knows an ExternalError is
Determinate, there are more opportunities to retry it. Critically, this
includes the "database is locked" error that can arise from concurrent
use of sqlite. This commit greatly reduces the number of ExternalErrors
the bubble up to WriteHandle::compare_and_append users when sqlite is
used as the Consensus impl.