persist: remove futures_executor::block_on from bench_compare_and_set
The next commit changes SqliteConsensus to run everything in
`tokio::task::block_in_place`, which only works with a real tokio
Runtime.

There's a bunch of other instances of this in the other old persist
benchmarks, but I've got a PR cooking that revamps all of these
micro-benchmarks, so I'm going to save those for another day.