feature-benchmark: Fix the KafkaRestart scenario
After recent refactoring the KafkaRestart scenario started failing
because the post-restart portion of the scenario was running
with the latest unstable Mz container rather than the version
that was under measurement.

Fix this by extending the scope of `c.override` to cover the entire
duration of the scenario. This way the `c.kill()` and `c.up()` actions
inside `RestartMz` operate on the correct image and restart Mz
using the right version.