doc: Transactional consistency design document
This commit adds a design doc for transactional consistency. It is in
service of correctness property #1, which is copied below:

Transactions in Materialize are strictly serializable with respect to
the operations that occur inside of Materialize. Operations include:
- SELECT, INSERT, UPDATE, and DELETE statements (but not TAIL)
- Materialize initiated acknowledgements of upstream sources (e.g., the
commit of an offset by a Kafka source, the acknowledge of an LSN by a
PostgresSQL source, etcetera)

However, this design document only focuses on the first bullet point,
and leaves the second bullet point for a follow up design document.

Works towards resolved #11631