persist: WIP use a connection pool in PostgresConsensus
Long-term, the connection limit will almost certainly be an issue, but
for now, this is the easiest way to get additional concurrency in the
persist Consensus impl used in development and CI. Hopefully, this will
be enough to get us through M1 and into M2.

Notably, this gets us good scaling up to the max pool size. We can do a
thing in 1.3ms and we can concurrently do the same thing 8X in a Runtime
with 8 thread and a pool with max 8 connections in 2.1ms. This should
help a lot with e.g. 22 system tables each advancing their frontier once
per second.

On a persist benchmarking machine

    MZ_PERSIST_RECORD_COUNT_SMALL=64 MZ_PERSIST_RECORD_SIZE_BYTES_SMALL=1024 MZ_PERSIST_BATCH_MAX_COUNT_SMALL=8
    consensus/compare_and_set/postgres/64KiB
                            time:   [1.3261 ms 1.3362 ms 1.3469 ms]
                            thrpt:  [46.404 MiB/s 46.776 MiB/s 47.131 MiB/s]
                     change:
                            time:   [-4.7604% +1.3972% +10.859%] (p = 0.77 > 0.05)
                            thrpt:  [-9.7957% -1.3780% +4.9983%]
                            No change in performance detected.
    consensus/concurrent_compare_and_set/postgres/64KiB
                            time:   [2.1304 ms 2.1728 ms 2.2393 ms]
                            thrpt:  [223.28 MiB/s 230.12 MiB/s 234.70 MiB/s]
                     change:
                            time:   [-84.453% -83.747% -83.012%] (p = 0.00 < 0.05)
                            thrpt:  [+488.65% +515.28% +543.20%]
                            Performance has improved.