persist: WIP use a connection pool in PostgresConsensus
Long-term, the connection limit will almost certainly be an issue, but
for now, this is the easiest way to get additional concurrency in the
persist Consensus impl used in development and CI. Hopefully, this will
be enough to get us through M1 and into M2.

Notably, this gets us good scaling up to the max pool size. We can do a
thing in 1.3ms and we can concurrently do the same thing 8X in a Runtime
with 8 thread and a pool with max 8 connections in 2.1ms. This should
help a lot with e.g. 22 system tables each advancing their frontier once
per second.

On a persist benchmarking machine

    MZ_PERSIST_RECORD_COUNT_SMALL=64 MZ_PERSIST_RECORD_SIZE_BYTES_SMALL=1024 MZ_PERSIST_BATCH_MAX_COUNT_SMALL=8
    consensus/compare_and_set/postgres/64KiB
                            time:   [1.4040 ms 1.4136 ms 1.4240 ms]
                            thrpt:  [43.891 MiB/s 44.213 MiB/s 44.515 MiB/s]
                     change:
                            time:   [-2.3115% +1.4342% +5.1769%] (p = 0.49 > 0.05)
                            thrpt:  [-4.9221% -1.4139% +2.3662%]
                            No change in performance detected.
    consensus/concurrent_compare_and_set/postgres/64KiB
                            time:   [2.1808 ms 2.2118 ms 2.2479 ms]
                            thrpt:  [222.43 MiB/s 226.06 MiB/s 229.28 MiB/s]
                     change:
                            time:   [-83.646% -83.100% -82.623%] (p = 0.00 < 0.05)
                            thrpt:  [+475.48% +491.71% +511.48%]
                            Performance has improved.