repr: Avoid some allocations for datetime parsing
The expected field for parsing pdt from tokens is equivalent to a peekable
iterator, which is more general than a mutable reference to a VecDeque.
Swap the types, and adjust all callsites to provide a reference to a type
that can be turned into an iterator.

This commit uses a trick that references to const declarations are 'static.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>