persist: use txns to implement PostgresConsensus
We originally did everything as single statements to save network
roundtrips to aurora. However, #12487 seems to be revealing that we're
not getting the ordering we expect out of Postgres (concretely, we seem
to be seeing stale results from `head`). Using transactions seem to help
this. We probably want to go back to something closer to the original
impl for production, but for M1 (and certainly for getting 12216
merged), this should be okay.

On a persist-benchmarking machine

    MZ_PERSIST_RECORD_COUNT_SMALL=64 MZ_PERSIST_RECORD_SIZE_BYTES_SMALL=1024 MZ_PERSIST_BATCH_MAX_COUNT_SMALL=8
    consensus/compare_and_set/postgres/64KiB
                            time:   [1.6464 ms 1.6597 ms 1.6806 ms]
                            thrpt:  [37.189 MiB/s 37.658 MiB/s 37.961 MiB/s]
    consensus/concurrent_compare_and_set/postgres/64KiB
                            time:   [15.539 ms 15.817 ms 16.212 ms]
                            thrpt:  [30.842 MiB/s 31.611 MiB/s 32.178 MiB/s]