computed,storaged,materialized: simplify and unify panic handling
Remove all of the custom panic handling code from materialized. It never
worked quite right anyway. In particular, the code to print backtraces
on SIGSEGV often both failed to print a backtrace *and* inhibited using GDB to
get a backtrace.

Instead, this commit changes our three core services to call
`ore::panic::set_abort_on_panic` and nothing more. This ensures that a
crash in any thread will bring down the entire process, which is a much
better default than limping along with a panicked thread, which is what
Rust does by default.

What we lose is the nice message asking for a bug report, but that's
not nearly as important in a cloud native world--plus that message
tended to get lost in the spew of log messages when computed, storaged,
and materialized all simultaneously crash.

Fix #12583.