storaged,computed,materialized: use stable UID/GID in Docker image
Use the stable 999 UID and GID for the "materialize" user in the
materialized, materialized-slim, storaged, and computed Docker images.
Having a stable UID/GID will prevent a painful migration later if the
Ubuntu base image changes how UIDs/GIDs are assigned; "999" seems to be
a loose Docker convention and is used by both the mysql and postgres
Docker images.

Note that using a less-privileged user is new behavior for the storaged
and computed images that bring it in line with the materialized and
materialized-slim images.