Use passed in AWS external ID as a prefix
Instead of using the external ID as passed in, use it as a prefix with
the source or sink ID appended.

This should play nicer with RBAC, in case customers want to allow only
certain sources or sinks to assume their AWS role.

Part 1 of https://github.com/MaterializeInc/cloud/issues/2837