testdrive: fail when expected errors don't match
Due to a bug in #12228, testdrive no longer fails when a SQL query does
not match its expected error. This commit fixes the bug.

The feature added in #12228--the ability to assert that a SQL query will
timeout--is preserved, but with new syntax. Rather than string matching
on a "deadline is exceeded" error message, a new `timeout` error
expectation is added. See the updated documentation within for details.