Introduce `SourceConnector::Log`
This commit introduces the `SourceConnector::Log` variant, which is
split from `SourceConnector::Local`. This split makes it possible to
differentiate log sources from other local sources.

The handling of the `SourceConnector::Log` is the same as that of
`SourceConnector::Local`. The only difference is that log sources are
listed with the "log" connector type in mz_sources:

```
materialize=> SELECT * FROM mz_sources LIMIT 1;
  id  |  oid  | schema_id |      name      | connector_type | volatility
------+-------+-----------+----------------+----------------+------------
 s245 | 20192 |         1 | mz_peek_active | log            | volatile
```