persist: make ColumnarRecords present T and D as `[u8; 8]`
This removes some TODOs in mz-persist-client to stop calling *_le_bytes.

Internally, the columnar format intentionally stores them both as i64
columns (as opposed to something like a fixed width binary column)
because this allows us additional compression options. Note that we use
an i64 for Time (currently a u64 in mz) because the useful range (nanos
since unix epoch) fits in [i64::MAX] until some time after year 2200.

This pushes a number of u64/i64 <-> [u8;8] conversions to other places,
because all the stuff in encoding.rs still hardcodes `Antichain<u64>`. I
took a swing at also fixing these up in this commit, but it quickly got
out of hand and most of the affected code is currently unused (e.g.
compaction). I think it'll be better to do this as part of the work to
make a permanent format for new persist.