sql: model REPLICA clause in CREATE CLUSTER as an option
Previously the `REPLICA` clause to `CREATE CLUSTER` had to appear first
and could not be followed by a comma. This was valid syntax:

    CREATE CLUSTER c
    REPLICA a (SIZE 'small'),
    REPLICA b (SIZE 'small')  -- note missing comma
    INTROSPECTION GRANULARITY '1s'

But this was not:

    CREATE CLUSTER c
    REPLICA a (SIZE 'small'),
    REPLICA b (SIZE 'small'), -- note comma
    INTROSPECTION GRANULARITY '1s'

This commit adjusts the syntax so that the latter is valid but the
former is not. This makes it more natural to intersperse replica
specifications with other options.

Touches #12660.