persist-client: opportunistically expire read handles on drop
Read handles must be expired when they are no longer needed in order to
release their requirements on compaction. Previously users had to
call the `ReadHandle::expire()` async destructor in all places where
a `ReadHandle` could go out of scope, but there were cases were this was
happening implicitly (e.g as part of dropping a closure and its state).

Since pretty much everything in materialize uses tokio this changes the
`Drop` implementation of `ReadHandle` to try and spawn an async task
that will expire the read handle as it goes out of scope.

The original behaviour is maintained when there is no tokio runtime
available.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>