persist: bring back auto expiry of ReadHandle on drop
This only works when a tokio Runtime is available in the TLC, but that
should be most/all places in mz.

Mechanically, this solves the Send+Sync problem that prevented us from
doing this originally by making the PhantomData in State use `fn() -> T`
types. According to the docs, PhantomData is to "mark things that act
like they own a T". We don't actually own K, V, or D, just the ability
to produce them. Using the `fn() -> T` pattern gets us the same variance
as T [1], but also allows State to correctly derive Send+Sync.

[1]: https://doc.rust-lang.org/nomicon/phantom-data.html#table-of-phantomdata-patterns

When possible, explicit expiry is still preferred because the Drop one
is best effort and is dependant on a tokio [Handle] being available in
the TLC at the time of drop (which is a bit subtle). Also, explicit
expiry allows for control over when it happens.