persist: also return Determinate errors from unreliable
The investigation of why maelstrom ops errored more than my mental model
said they should revealed two sources of mismatch. The first was that
unreliable always returned (indeterminate) timeout errors, which was a
bug that this commit fixes. (The other was that my mental model didn't
account for CaS failures when the state was out of date.)

Also included is some debugging changes that were helpful in tracing
this down. Finally, there are also a couple small fixes for ways this
has rotted given that I haven't gotten a chance to hook it up to CI yet.