ore/tracing: only prefix logs with service name in process orchestrator
The `--log-process-name` argument was introduced so that log lines would
only be prefixed with the process name when using the process
orchestrator, which intermingles log output from all services. But this
never actually got wired up, and the `--log-process-name` option is
enabled in both the process and Kubernetes orchestrator.

This commit does that wiring. The `--log-process-name` option turns out
to be unnecessary; the presence of the `--pid-file-location` can be used
to determine when a service is running under the process orchestrator
and therefore the `log_service_name` tracing configuration option should
be enabled.