orchestrator: plumb tracing configuration to services
This commit allows tracing configuration to be easily plumbed through
all services created by an orchestrator.

The bulk of the work is done by a new orchestrator implementation called
`TracingOrchestrator`. This wraps an existing orchestrator and injects
tracing arguments into all services created by the inner orchestrator.
It also provides a command-line parser for the tracing options that can
be used in all tracing-aware services.

Co-authored-by: Gus Wynn <gus@materialize.com>