orchestrator: improve argument generation
Introduce a `ServiceAssignments` type that is passed to the closure that
generates the arguments for a service (i.e., in the
`ServiceConfig::args` field).

This greatly simplifies the type signature for the closure and
introduces a more natural place for documentating what each of the
parameters means.

This commit also folds `Orchestrator::listen_host` into
`ServiceAssignments`. This will become important in a future commit
where services are created only access to a `NamespacedOrchestrator` and
not an `Orchestrator`.