compute: hoist kubernetes ordinal hack into Dockerfile
Kubernetes does not directly support telling a replica of a
`StatefulSet` which ordinal index it is in the set. This is a problem
for `computed`, which needs to tell timely which process index it is in
the cluster. As a hack, in Kubernetes, we can parse this information out
of the `HOSTNAME` field.

This commit hoists that parsing out of the `computed` binary itself and
into the Dockerfile, as requested in the code review of #12086. This
localizes this hacky logic a bit better.

Additionally, this commit takes the opportunity to simplify the CLI
arguments to `computed`. The number of processes is implicit in the
number of provided hostnames, so there is no need for a separate
`--processes` argument. Similarly, the address file is never used
either, so this commit remove it.