storage: move persist since downgrades outside of main coordinator task
This adds a wrapper that holds on to backing persist shards/collections
that the storage controller is aware of. The handles hold back the since
frontier and we need to downgrade them when the read capabilities
change.

Internally, this has an async task and the methods for registering a
handle and downgrading sinces add commands to a queue that this task is
working off. This makes the methods non-blocking and moves the work
outside the main coordinator task, meaning the coordinator is spending
less time waiting on persist calls.

We can move `downgrade_since()` calls out of the main task because the
caller doesn't depend on the result, nor is sequencing important with
these calls. With `append()` things are not as simple because the caller
(usually Coordinator) needs to know that a call succeeded before it
continues with other (potentially followup) actions. We could think
about just moving the periodic "advance frontier" calls to a separate
task and not use `append()` with empty updates for that, though.