platform-upgrade: Minimum viable subset of the framework
The new platform upgrade framework separates the individual features
to be checked on an upgrade into self-contained Check classes that
contain both the population and the validation logic for each feature.

The Checks will then be run at the appropriate times during an
upgrade Scenario. This commit provides a dummy Scenario, where we
are not upgrading or restarting anything but still perform the
population and validation steps.