persist: make info logging less spammy
A number of these happen often enough that the show up regularly in the
logs. We've received the question of whether they are ignorable a number
of times (plus it's spammy), so downgrade the first few attempts through
each of the retry loops to debug!.

I tried to use `tracing::event` and pass it a tracing::Level instead of
duplicating the log message in both halves of an if branch, but it
complained. Knowing the impl of tracing, this makes sense, but it's a
touch unfortunate. I wonder if there's still some way to accomplish the
same thing without so much duplication.