mzbuild: allow specifying architecture to build for
Teach the `mzimage` command to accept the `--arch` flag, as in

    $ mzimage acquire --arch=x86_64 materialized

which allows cross-compiling for an architecture other than the host.
This works wonderfully on macOS where we have cross-compiling toolchains
that run on the host. On Linux it is too slow to be useful, though, as
Docker will attempt to compile Rust via QEMU emulation.