sql[-parser]: integrate connectors with name resolution
Connector references in `CREATE SOURCE` statements, as in

    CREATE SOURCE ... FROM KAFKA CONNECTOR db.schema.connector

were currently handled with custom name resolution and dependency
tracking code. But they work just like table references in a `SELECT`
statement.

This commit pivots connectors to use the standard name resolution
infrastructure, which eliminates a heap of complexity.

Part of the connector epic (#11504).