sql-parser: move `Raw*` types out of ast::defs module
The `Raw*` types (e.g., `RawObjectName`) were incorrectly getting
included in the generated `Fold`, `Visit`, and `VisitMut` traits. These
types are only conditionally part of the AST (when an AST node is
instantiated with `<Raw>`), so move them out of the ast::defs module.

Also rename `UnresolvedDataType` to `RawDataType` to indicate that it
only ever appears in the AST via `Raw::DataType`.