test: move expected values of equality checks in the results
Our testdrive suite has a large number of equality checks that are
evaluated in the SQL context and the testdrive framework only verifies
that the equality was true.

When a test like this fails it's often useful to also know the value
that didn't match the equality check.

By moving the expected values in the results when a test fails the
developer will be informed of both the expected and the actual value of
the equality.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>