*: plumb PersistClientCache to everywhere it's needed
We need a singleton PersistClientCache so that it can internally share
postgres/aurora connections. It also needs to have access to a
MetricsRegistry so persist can internally hook up its metrics, which
this static singleton workaround introduced in 12216 doesn't. So,
replace the workaround with instantiating it once at storaged/computed
startup and explictly plumbing it everywhere it needs to go.