storage: s3: remove `select!` macros
Following the saga of eliminating `select!` calls from the codebase,
these two particular cases don't even need a `select!` because there is
just one branch that does work in a loop and another branch that
receives a shutdown notification, cancelling the other branch.

In order to avoid cancelling the futuring doing the work at some random
point we must explicitly put a point where we deem is appropriate to
exit and cleanup. However, if we do so we can directly use the shutdown
watch receiver without concurrent polling.

This has the effect that a shutdown notification won't apply immediately
and it will wait for one more message to arrive to the source reader but
that's unavoidable if we want to not cancel the future.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>